Refactor: Update AlarmManagerScreenModel to handle Int UiEvents
Changed the `AlarmManagerScreenModel` to handle `UiEvent<Int>` instead of
 `UiEvent<Long>`. This likely reflects a change in the type of data being passed through UI events related to alarm management.

- Modified `sendUiEvent` to accept `UiEvent<Int>`.
- Updated the type parameter for `UiEventHandler` to `Int`.

+++ services/notifications/src/main/java/city/zouitel/notifications/viewmodel/AlarmManagerScreenModel.kt
@@ -21,7 +21,7 @@ import kotlinx.coroutines.launch
 class AlarmManagerScreenModel(
     private val context: Context,
     private val updateReminder: ReminderUseCase.Update
-): ScreenModel, UiEventHandler<Long> {
+): ScreenModel, UiEventHandler<Int> {
 
     private val intent = Intent(context , NotificationReceiver::class.java)
     private val alarmManager = context.getSystemService(Context.ALARM_SERVICE) as AlarmManager
@@ -68,7 +68,7 @@ class AlarmManagerScreenModel(
         )
     }
 
-    override fun sendUiEvent(event: UiEvent<Long>) {
+    override fun sendUiEvent(event: UiEvent<Int>) {
         when(event) {
             is UiEvent.Update -> performUiEvent { updateReminder(event.data) }
             else -> throw NotImplementedError("UiEvent not implemented: $event")
